backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.25.0
pydantic>=2.6.0
pymongo>=4.5.0
python-dotenv>=1.0.0
httpx>=0.25.0
google-auth>=2.23.0
python-multipart>=0.0.6
numpy>=1.26.0